Asynchrones Hosting recherche
Research associated to asynchrony hosting
 --> what is it? how does it work? what is the difference between synchrony hosting?

In addition there are several applications, which can run on this format.

On the other hand, most of the asyn. programs are in C written and the http environment is implemented with js.

My Opinion:
The implementation of the asyn. programs is quite difficult and complex, but will be further analyzed. A few general aspects of asyn. is already known and the base function of it, is also not so complicated. So considering these points, it is definitely possible to implement in our environment.