Asyn_hosting
First i deleted the 2 libraries, because copy-right etc.

Then i made a project under arduiono ide with a readme.txt

👽
The readme.txt contains the licences of the libraries i used

The main code "async" contains only a base structur, which methods, which will be modified in the future.

But in addition the string_data, which will contain all information about the sensors, was divided in 4 sections "4 sensors data". With that method the graphical way of the website, will be much smoother and easier to implement.

Ofc, some code parts were not known, so i searched those in the internet and looked up some tutorials. On zotero you can see my search history.