changes
i made a test program if we check the time with js and give it back to the esp.

for that i used Date, which works with the UTC time zone.  The time zone is displayed on a label, which updates with every event or http request.

The html only shows the hpp sensor data. In addition this only raw code, meaning i was not able to test it. Mayer has to test via his esp, to see if he can really extract the date, which should work, but we'll see.

Further more a added a folder named, Website_html_js --> this will be the actual website, which will be modified.