changes
first  i changed a few folder names, cause they were not appropriate.

then i started with the website (i also changed the location of the folder, it's important yk), for the website i started changing the html (the important looking part), added a css, js file.

The js file has already his functions (but empty), the css is almost empty. But the base design of the html is finished. There were a few problems with stroke-width, cause its not working, idk why, it should work, but its not, nothing to horrible to worry about, but it still annoys me.

So in summarize the html is almost finished, and tomorrow i will set my sights to the js.file and implement the xy graph on the canvas.