changes code
First I i added a graph. This graph (x,y) will show the sensor data. This was accomplished with the LineTo, where I could draw Lines, to create a graph.

In addition I added a read method for the comboxes. Where the program can get the actual value of the comboxes. It is very important later, for the graph.

Further more, i could not found a method for the border, i think that it is not possible to set borders for a canvas. That are my thoughts.