html canvas (graph) x axe changes
i updated the graph today. There are several changes. First there is a reset command for the graph, so the graph can have a reset, when the function (e.g. canvas_setting()) is being applied. But it deletes a few lines, this problem will be fixed tomorrow.

Further more, the time zones were set, meaning, the graph can set his x axe to hours, days. This means our graph can show the user pieces of information of our sensor from  a whole year.