changes: canvas x axe bugs fixed and naming of the x axe
First I fixed the problems we had till now, the first one was the clearing method. It cleared way too much or way too less. The fixation was --> changing the coordinates for the method. That was all, what was needed to fix the bug.  The coordinates had to be changed, to the line width, in our case it, our line width was 2, so i had to change the method ctx.clearRect(x,1,x,x) to that. And of course i had to do it again, for width/height.

Then I added the naming of the x axe. At first it was problematic to add a text in a canvas, because I never did that in js. But after i managed to add it, we could show the timelines, with the correct named x axe. But the texts added on each other, which looked very ugly. So i had to write a method, to fix that problem.

That is all for today