Changes graph and http request
-First http request was fixed via if
-There were a few graph changes
First the graph is almost completed, there is still a small bug will deleting the lapse names. But this will be fixed tomorrow (it is a small problem)
So, what changed? The graph's x,y are named fully, the scaling has been done, so that it will to graph width and heigth aumatically configure its lines. The combobox problem was also fixed, where the value was not directly used, i packed in canvas_setting so it will definitely use it.