bug fix of the graph (x-axe)
It is a small change, but an important change. I fixed the appearance of double symbols when changing the time zone, like from "today" to "this month", etc. The symbols were overlapping, which should not be possible. The problem was, that the, clearRect could not reach the area, where these overlappings were happening and that of course could not solve the issue. So i fixed that and expanded his range.

A small reminder, the whole program, is written in a format, so that it can automatically adjust to the screen size. But of course, i did not test that yet, i have to test it someday. Theoretically (code base) it is there and should of course work, you never know.

That is my report about today.